move test modules and snaps to appropriate folders
